opcode: Add dialect :io for I/O Revision II.
I/O Revision II uses a dialect that deviates a lot from the standard dialect. Both changes to existing opcodes and additions of extra opcodes are present.

This is still WIP and only a small number of opcodes are added/updated currently. Future work may include adding support for all the rest of the opcodes and cross-verify with decompiled interpreter code.